Restructured matchCard file in components folder.
